+++ app-api/src/test/java/com/tasteam/infra/messagequeue/NotificationMessageQueueFlowIntegrationTest.java
@@ -12,12 +12,13 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Mockito;
+import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.json.JsonMapper;
 import com.tasteam.config.annotation.MessageQueueFlowTest;
 import com.tasteam.domain.group.event.GroupMemberJoinedEvent;
 import com.tasteam.domain.notification.entity.NotificationType;
@@ -26,7 +27,7 @@ import com.tasteam.domain.notification.service.NotificationService;
 import jakarta.annotation.Resource;
 
 @MessageQueueFlowTest
-@Import(NotificationMessageQueueFlowIntegrationTest.TestConfig.class)
+@SpringBootTest(classes = NotificationMessageQueueFlowIntegrationTest.TestConfig.class, webEnvironment = SpringBootTest.WebEnvironment.NONE)
 @DisplayName("[통합](Notification) NotificationMessageQueueFlow 통합 테스트")
 class NotificationMessageQueueFlowIntegrationTest {
 
@@ -114,6 +115,30 @@ class NotificationMessageQueueFlowIntegrationTest {
 
 	@Configuration
 	static class TestConfig {
+
+		@Bean
+		MessageQueueProperties messageQueueProperties() {
+			MessageQueueProperties properties = new MessageQueueProperties();
+			properties.setEnabled(true);
+			properties.setProvider(MessageQueueProviderType.REDIS_STREAM.value());
+			properties.setDefaultConsumerGroup("tasteam-api");
+			return properties;
+		}
+
+		@Bean
+		ObjectMapper objectMapper() {
+			return JsonMapper.builder().findAndAddModules().build();
+		}
+
+		@Bean
+		MessageQueueProducer messageQueueProducer() {
+			return Mockito.mock(MessageQueueProducer.class);
+		}
+
+		@Bean
+		MessageQueueConsumer messageQueueConsumer() {
+			return Mockito.mock(MessageQueueConsumer.class);
+		}
 
 		@Bean
 		NotificationService notificationService() {
